requirements.txt
anthropic
pandas
numpy
dbt-duckdb
python-dotenv
